+++ lib/widgets/ny_page.dart
@@ -51,6 +51,7 @@ abstract class NyPage<T extends StatefulWidget> extends NyBaseState<T> {
         if (eventHistory.isNotEmpty) {
           stateData = eventHistory.last.event.props[1];
         }
+
         eventSubscription = eventBus!.on<UpdateState>().listen((event) async {
           if (event.stateName != stateName) return;
 
@@ -151,8 +152,24 @@ abstract class NyPage<T extends StatefulWidget> extends NyBaseState<T> {
     }
 
     String action = data['action'];
+    dynamic actionData = data.containsKey('data') ? data['data'] : null;
+
     if (stateActions.containsKey(action)) {
-      await stateActions[action]!();
+      final function = stateActions[action]!;
+
+      String functionString = function.runtimeType.toString();
+
+      // Determine if the function takes parameters based on its toString representation
+      bool hasParameters = functionString.contains("(dynamic)") ||
+          functionString.contains("(Object?)") ||
+          !functionString.contains("()");
+
+      if (hasParameters) {
+        await Function.apply(function, [actionData]);
+        return;
+      }
+
+      await Function.apply(function, []);
     }
   }
 }